+++ src/dipdup/datasources/tzkt/datasource.py
@@ -1059,6 +1059,7 @@ class TzktDatasource(IndexDatasource):
             originated_contract_address=originated_contract_json.get('address'),
             originated_contract_type_hash=originated_contract_json.get('typeHash'),
             originated_contract_code_hash=originated_contract_json.get('codeHash'),
+            originated_contract_tzips=originated_contract_json.get('tzips'),
             storage=operation_json.get('storage'),
             diffs=operation_json.get('diffs') or (),
             delegate_address=delegate_json.get('address'),