+++ CHANGELOG.md
@@ -6,6 +6,11 @@
 
 * tzkt: Added `originated_contract_tzips` field to `OperationData`.
 
+### Fixed
+
+* jobs: Fixed jobs with `daemon` schedule never start.
+* jobs: Fixed failed jobs not throwing exceptions into the main loop.
+
 ## 5.1.1 - 2022-05-13
 
 ### Fixed